pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pystages"
version = "1.1.1"
authors = [
  { name="Olivier Hérivaux", email="olivier.heriveaux@ledger.fr" },
  { name="Michaël Mouchous", email="michael.mouchous@ledger.fr" },
]
description = "Motorized stage control library for scientific applications"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial",
    "pyusb",
    "numpy",
    "pyqt6"
]

[project.urls]
"Homepage" = "https://github.com/Ledger-Donjon/pystages"
"Bug Tracker" = "https://github.com/Ledger-Donjon/pystages/issues"
"Documentation" = "https://pystages.readthedocs.io/en/latest"
